apps/api/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
gunicorn>=22.0.0
SQLAlchemy>=2.0.30
asyncpg>=0.29.0
pgvector>=0.2.5
pydantic>=2.7.0
pydantic-settings>=2.2.1
orjson>=3.10.0
httpx>=0.27.0
tenacity>=8.3.0
beautifulsoup4>=4.12.0
python-dateutil>=2.9.0
jinja2>=3.1.4
python-multipart>=0.0.9
apscheduler>=3.10.4
numpy>=1.26.0
umap-learn>=0.5.6
scikit-learn>=1.4.0
